requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
stagehand-py>=0.3.6
playwright>=1.35.0
openai==1.26.0
pydantic==2.6.4
python-dotenv==1.0.0
httpx>=0.24.0
requests>=2.31.0
